Added separate life display
Implemented a separate animation that displays only the lifes and added a transition check that once it's done it goes back to being hidden. I've also added in a check that once the player wants to display the entire HUD while only lifes is being displayed it won't hide it first but rather displays the remaining UI elements.