Improved CheckIfRunning
Noticed that when performing the bodyslam the player got stuck in mid-air. To fix this I've build in an extra check to see if the player is performing the bodyslam in the CheckIfRunning function. If so the players movement gets update so he won't stay stuck in mid-air. Also cleaned up the code in CheckIfRunning to be more effecient.