Improved ResetCrates function
Inside the Game Manager script In the ResetCrates function I've done an extra check to see if the IsResetGame is true. If so then we set it to false and loop over all the crates in TotalBrokenCrates and call the function ResetCrate to reset all the crates back to their original state.