Overhauled Player script
After discussing with a fellow Unity programmer on how I used the to have my playerscript setup we both agreed that it was to complicated and way too messy. So I've collabed with him to come up with a better playerscript and making use of the RigidBody component for the movement. This turned out as a way better version that what I had before with a lot of complications that I would've come across in the future solved. This is how I'm going to use my playerscript for the future of this project.