Display Monitor Added
Went online looking for a monitor I could use in my game to display how many crates have been broken and what the total amount of crates is. Eventually I found one and implemented this into my game. Then Used TexMeshPro to display the text in front of the monitor es if it was being displayed by the monitor. I've also added the TextMeshPro text to the Game Manager script so it would get updated with crate count as well.