Fixed barrier behaviour
Inside Game Manager I've made a reference to the BarrierManager. Inside the ResetCrates function at the end I check if the CheckCrateTotal is false. If so, then I call the function EnableLasers on the barrier manager script to close the laser barrier again so they won't stay open after the reset.