Improved Companion robot script
Instead of checking it's own Target is equal to null I should be check if the companion position object doesn't already have a companion robot as it's child. If it doesn't we call the function PositionRobot and return the function. Otherwise we increase the extra hit count and call the function DestroyObject.